changes as per comments
