+++ openmetadata-ui/src/main/resources/ui/playwright/constant/common.ts
@@ -1,6 +1,3 @@
-import { PolicyRulesType } from '../support/access-control/PoliciesClass';
-import { GlobalSettingOptions } from './settings';
-
 /*
  *  Copyright 2024 Collate.
  *  Licensed under the Apache License, Version 2.0 (the "License");
@@ -33,128 +30,3 @@ export const NAME_MIN_MAX_LENGTH_VALIDATION_ERROR =
 
 export const NAME_MAX_LENGTH_VALIDATION_ERROR =
   'Name size must be between 1 and 128';
-
-export const DATA_STEWARD_RULES: PolicyRulesType[] = [
-  {
-    name: 'DataStewardRole',
-    resources: ['All'],
-    operations: [
-      'EditDescription',
-      'EditDisplayName',
-      'EditLineage',
-      'EditOwners',
-      'EditTags',
-      'ViewAll',
-    ],
-    effect: 'allow',
-  },
-];
-
-export const GLOBAL_SETTING_PERMISSIONS: Record<
-  string,
-  { testid: GlobalSettingOptions; isCustomProperty?: boolean }
-> = {
-  metadata: {
-    testid: GlobalSettingOptions.METADATA,
-  },
-  customAttributesDatabase: {
-    testid: GlobalSettingOptions.DATABASES,
-    isCustomProperty: true,
-  },
-  customAttributesDatabaseSchema: {
-    testid: GlobalSettingOptions.DATABASE_SCHEMA,
-    isCustomProperty: true,
-  },
-  customAttributesStoredProcedure: {
-    testid: GlobalSettingOptions.STORED_PROCEDURES,
-    isCustomProperty: true,
-  },
-  customAttributesTable: {
-    testid: GlobalSettingOptions.TABLES,
-    isCustomProperty: true,
-  },
-  customAttributesTopics: {
-    testid: GlobalSettingOptions.TOPICS,
-    isCustomProperty: true,
-  },
-  customAttributesDashboards: {
-    testid: GlobalSettingOptions.DASHBOARDS,
-    isCustomProperty: true,
-  },
-  customAttributesPipelines: {
-    testid: GlobalSettingOptions.PIPELINES,
-    isCustomProperty: true,
-  },
-  customAttributesMlModels: {
-    testid: GlobalSettingOptions.MLMODELS,
-    isCustomProperty: true,
-  },
-  customAttributesSearchIndex: {
-    testid: GlobalSettingOptions.SEARCH_INDEXES,
-    isCustomProperty: true,
-  },
-  customAttributesGlossaryTerm: {
-    testid: GlobalSettingOptions.GLOSSARY_TERM,
-    isCustomProperty: true,
-  },
-  customAttributesAPICollection: {
-    testid: GlobalSettingOptions.API_COLLECTIONS,
-    isCustomProperty: true,
-  },
-  customAttributesAPIEndpoint: {
-    testid: GlobalSettingOptions.API_ENDPOINTS,
-    isCustomProperty: true,
-  },
-  bots: {
-    testid: GlobalSettingOptions.BOTS,
-  },
-};
-export const ID: Record<
-  string,
-  { testid: GlobalSettingOptions; button: string; api?: string }
-> = {
-  teams: {
-    testid: GlobalSettingOptions.TEAMS,
-    button: 'add-team',
-  },
-  users: {
-    testid: GlobalSettingOptions.USERS,
-    button: 'add-user',
-    api: '/api/v1/users?*',
-  },
-  admins: {
-    testid: GlobalSettingOptions.ADMINS,
-    button: 'add-user',
-    api: '/api/v1/users?*',
-  },
-  databases: {
-    testid: GlobalSettingOptions.DATABASES,
-    button: 'add-service-button',
-    api: '/api/v1/services/databaseServices?*',
-  },
-  messaging: {
-    testid: GlobalSettingOptions.MESSAGING,
-    button: 'add-service-button',
-    api: '/api/v1/services/messagingServices?*',
-  },
-  dashboard: {
-    testid: GlobalSettingOptions.DASHBOARDS,
-    button: 'add-service-button',
-    api: '/api/v1/services/dashboardServices?*',
-  },
-  pipelines: {
-    testid: GlobalSettingOptions.PIPELINES,
-    button: 'add-service-button',
-    api: '/api/v1/services/pipelineServices?*',
-  },
-  mlmodels: {
-    testid: GlobalSettingOptions.MLMODELS,
-    button: 'add-service-button',
-    api: '/api/v1/services/mlmodelServices?*',
-  },
-  storage: {
-    testid: GlobalSettingOptions.STORAGES,
-    button: 'add-service-button',
-    api: '/api/v1/services/storageServices?*',
-  },
-};
